test(daemon): move linux_watchdog tests out of the source file
`make check-daemon` forbids inline test code in crates/*/src outside a
tests/ directory (.agents/testing.md). Move the watchdog open-error
classification tests into src/tests/linux_watchdog.rs and register the
module in src/tests/mod.rs; make OpenFailure + classify_open_error
pub(crate) so the external test module can reach them.
